fix(ci): _snprintf alias on POSIX, correct testdata path, wsprintfW va_arg (parity)
- posix/windows.h maps msvcrt _snprintf/_vsnprintf to C99 snprintf/vsnprintf
  (trace.c uses _snprintf; there is no _snprintf on Linux/macOS -> link error).
- parity.yml: testdata lives at tools/wine_usp/testdata, i.e. ../testdata from
  port/ (not ../../testdata); compare_trace.py is at tools/wine_usp/ too.
- posix_compat.c wsprintfW: %x/%X/%o read a va_arg unsigned int (was declared
  unsigned long -> format warnings).
Local -fshort-wchar parity still 161/161.

+++ tools/wine_usp/port/posix/posix_compat.c
@@ -147,8 +147,8 @@ int WINAPI wsprintfW(LPWSTR out, LPCWSTR fmt, ...)
             const char *f = (p[1]==L'd')?"%ld":(p[1]==L'u')?"%lu":(p[1]==L'X')?"%lX":"%lx";
             snprintf(tmp, sizeof(tmp), f, v); d += emit_ascii(d, tmp); p += 2; continue;
         }
-        if (*p == L'x' || *p == L'X') { unsigned long v = va_arg(ap, unsigned int); snprintf(tmp, sizeof(tmp), (*p=='X')?"%X":"%x", v); d += emit_ascii(d, tmp); p++; continue; }
-        if (*p == L'o') { unsigned long v = va_arg(ap, unsigned int); snprintf(tmp, sizeof(tmp), "%o", v); d += emit_ascii(d, tmp); p++; continue; }
+        if (*p == L'x' || *p == L'X') { unsigned int v = (unsigned int)va_arg(ap, unsigned int); snprintf(tmp, sizeof(tmp), (*p=='X')?"%X":"%x", v); d += emit_ascii(d, tmp); p++; continue; }
+        if (*p == L'o') { unsigned int v = (unsigned int)va_arg(ap, unsigned int); snprintf(tmp, sizeof(tmp), "%o", v); d += emit_ascii(d, tmp); p++; continue; }
         *d++ = L'%'; if (*p) *d++ = *p++;
     }
     *d = 0;
